Load a model from checkpoint
Play with the model. But using MCTS is not a great way to evaluate the model -- MCTS is all you need for mastery of TicTacToe.

I want to know how good the model is by itself, without MCTS being used. It should still be able to play decent TicTacToe that way.